Add protocol name&version to net attribute getters
