+++ instrumentation-api-semconv/src/main/java/io/opentelemetry/instrumentation/api/instrumenter/messaging/MessagingAttributesExtractor.java
@@ -81,11 +81,6 @@ public final class MessagingAttributesExtractor<REQUEST, RESPONSE>
           SemanticAttributes.MESSAGING_DESTINATION_NAME,
           getter.getDestination(request));
     }
-    internalSet(attributes, SemanticAttributes.NET_APP_PROTOCOL_NAME, getter.getProtocol(request));
-    internalSet(
-        attributes,
-        SemanticAttributes.NET_APP_PROTOCOL_VERSION,
-        getter.getProtocolVersion(request));
     internalSet(
         attributes,
         SemanticAttributes.MESSAGING_MESSAGE_CONVERSATION_ID,
