+++ instrumentation-api-semconv/src/main/java/io/opentelemetry/instrumentation/api/instrumenter/net/internal/InternalNetServerAttributesExtractor.java
@@ -32,7 +32,14 @@ public final class InternalNetServerAttributesExtractor<REQUEST> {
   }
 
   public void onStart(AttributesBuilder attributes, REQUEST request) {
+
     internalSet(attributes, SemanticAttributes.NET_TRANSPORT, getter.getTransport(request));
+    internalSet(
+        attributes, SemanticAttributes.NET_APP_PROTOCOL_NAME, getter.getProtocolName(request));
+    internalSet(
+        attributes,
+        SemanticAttributes.NET_APP_PROTOCOL_VERSION,
+        getter.getProtocolVersion(request));
 
     boolean setSockFamily = false;
