+++ instrumentation/apache-pulsar/apache-pulsar-2.8/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/pulsar/v2_8/telemetry/PulsarMessagingAttributesGetter.java
@@ -37,18 +37,6 @@ enum PulsarMessagingAttributesGetter implements MessagingAttributesGetter<Messag
 
   @Nullable
   @Override
-  public String getProtocol(Message<?> message) {
-    return null;
-  }
-
-  @Nullable
-  @Override
-  public String getProtocolVersion(Message<?> message) {
-    return null;
-  }
-
-  @Nullable
-  @Override
   public String getConversationId(Message<?> message) {
     return null;
   }
