+++ instrumentation/jms/jms-common/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jms/JmsMessageAttributesGetter.java
@@ -42,18 +42,6 @@ enum JmsMessageAttributesGetter implements MessagingAttributesGetter<MessageWith
 
   @Nullable
   @Override
-  public String getProtocol(MessageWithDestination messageWithDestination) {
-    return null;
-  }
-
-  @Nullable
-  @Override
-  public String getProtocolVersion(MessageWithDestination messageWithDestination) {
-    return null;
-  }
-
-  @Nullable
-  @Override
   public String getConversationId(MessageWithDestination messageWithDestination) {
     try {
       return messageWithDestination.message().getJmsCorrelationId();
