+++ instrumentation/kafka/kafka-clients/kafka-clients-common/library/src/main/java/io/opentelemetry/instrumentation/kafka/internal/KafkaConsumerAttributesGetter.java
@@ -38,18 +38,6 @@ enum KafkaConsumerAttributesGetter implements MessagingAttributesGetter<KafkaPro
 
   @Override
   @Nullable
-  public String getProtocol(KafkaProcessRequest request) {
-    return null;
-  }
-
-  @Override
-  @Nullable
-  public String getProtocolVersion(KafkaProcessRequest request) {
-    return null;
-  }
-
-  @Override
-  @Nullable
   public String getConversationId(KafkaProcessRequest request) {
     return null;
   }
