+++ instrumentation/kafka/kafka-clients/kafka-clients-common/library/src/main/java/io/opentelemetry/instrumentation/kafka/internal/KafkaProducerAttributesGetter.java
@@ -44,18 +44,6 @@ enum KafkaProducerAttributesGetter
 
   @Override
   @Nullable
-  public String getProtocol(KafkaProducerRequest request) {
-    return null;
-  }
-
-  @Override
-  @Nullable
-  public String getProtocolVersion(KafkaProducerRequest request) {
-    return null;
-  }
-
-  @Override
-  @Nullable
   public String getConversationId(KafkaProducerRequest request) {
     return null;
   }
