+++ instrumentation/kafka/kafka-clients/kafka-clients-common/library/src/main/java/io/opentelemetry/instrumentation/kafka/internal/KafkaReceiveAttributesGetter.java
@@ -46,18 +46,6 @@ enum KafkaReceiveAttributesGetter implements MessagingAttributesGetter<KafkaRece
 
   @Override
   @Nullable
-  public String getProtocol(KafkaReceiveRequest request) {
-    return null;
-  }
-
-  @Override
-  @Nullable
-  public String getProtocolVersion(KafkaReceiveRequest request) {
-    return null;
-  }
-
-  @Override
-  @Nullable
   public String getConversationId(KafkaReceiveRequest request) {
     return null;
   }
