+++ instrumentation/rabbitmq-2.7/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/rabbitmq/RabbitChannelAttributesGetter.java
@@ -37,18 +37,6 @@ enum RabbitChannelAttributesGetter implements MessagingAttributesGetter<ChannelA
 
   @Nullable
   @Override
-  public String getProtocol(ChannelAndMethod channelAndMethod) {
-    return null;
-  }
-
-  @Nullable
-  @Override
-  public String getProtocolVersion(ChannelAndMethod channelAndMethod) {
-    return null;
-  }
-
-  @Nullable
-  @Override
   public String getConversationId(ChannelAndMethod channelAndMethod) {
     return null;
   }
