+++ instrumentation/rabbitmq-2.7/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/rabbitmq/RabbitDeliveryAttributesGetter.java
@@ -45,18 +45,6 @@ enum RabbitDeliveryAttributesGetter implements MessagingAttributesGetter<Deliver
 
   @Nullable
   @Override
-  public String getProtocol(DeliveryRequest request) {
-    return null;
-  }
-
-  @Nullable
-  @Override
-  public String getProtocolVersion(DeliveryRequest request) {
-    return null;
-  }
-
-  @Nullable
-  @Override
   public String getConversationId(DeliveryRequest request) {
     return null;
   }
