+++ instrumentation/rabbitmq-2.7/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/rabbitmq/RabbitReceiveAttributesGetter.java
@@ -47,18 +47,6 @@ enum RabbitReceiveAttributesGetter
 
   @Nullable
   @Override
-  public String getProtocol(ReceiveRequest request) {
-    return null;
-  }
-
-  @Nullable
-  @Override
-  public String getProtocolVersion(ReceiveRequest request) {
-    return null;
-  }
-
-  @Nullable
-  @Override
   public String getConversationId(ReceiveRequest request) {
     return null;
   }
