+++ instrumentation/rocketmq/rocketmq-client/rocketmq-client-4.8/library/src/main/java/io/opentelemetry/instrumentation/rocketmqclient/v4_8/RocketMqConsumerAttributeGetter.java
@@ -37,18 +37,6 @@ enum RocketMqConsumerAttributeGetter implements MessagingAttributesGetter<Messag
 
   @Nullable
   @Override
-  public String getProtocol(MessageExt request) {
-    return null;
-  }
-
-  @Nullable
-  @Override
-  public String getProtocolVersion(MessageExt request) {
-    return null;
-  }
-
-  @Nullable
-  @Override
   public String getConversationId(MessageExt request) {
     return null;
   }
