+++ instrumentation/rocketmq/rocketmq-client/rocketmq-client-4.8/library/src/main/java/io/opentelemetry/instrumentation/rocketmqclient/v4_8/RocketMqProducerAttributeGetter.java
@@ -42,18 +42,6 @@ enum RocketMqProducerAttributeGetter
 
   @Nullable
   @Override
-  public String getProtocol(SendMessageContext sendMessageContext) {
-    return null;
-  }
-
-  @Nullable
-  @Override
-  public String getProtocolVersion(SendMessageContext sendMessageContext) {
-    return null;
-  }
-
-  @Nullable
-  @Override
   public String getConversationId(SendMessageContext sendMessageContext) {
     return null;
   }
