+++ instrumentation/rocketmq/rocketmq-client/rocketmq-client-5.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/rocketmqclient/v5_0/RocketMqConsumerProcessAttributeGetter.java
@@ -42,18 +42,6 @@ enum RocketMqConsumerProcessAttributeGetter
 
   @Nullable
   @Override
-  public String getProtocol(MessageView messageView) {
-    return null;
-  }
-
-  @Nullable
-  @Override
-  public String getProtocolVersion(MessageView messageView) {
-    return null;
-  }
-
-  @Nullable
-  @Override
   public String getConversationId(MessageView messageView) {
     return null;
   }
