+++ instrumentation/rocketmq/rocketmq-client/rocketmq-client-5.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/rocketmqclient/v5_0/RocketMqProducerAttributeGetter.java
@@ -42,18 +42,6 @@ enum RocketMqProducerAttributeGetter
 
   @Nullable
   @Override
-  public String getProtocol(PublishingMessageImpl message) {
-    return null;
-  }
-
-  @Nullable
-  @Override
-  public String getProtocolVersion(PublishingMessageImpl message) {
-    return null;
-  }
-
-  @Nullable
-  @Override
   public String getConversationId(PublishingMessageImpl message) {
     return null;
   }
