+++ instrumentation/spring/spring-integration-4.1/library/src/main/java/io/opentelemetry/instrumentation/spring/integration/v4_1/SpringMessagingAttributesGetter.java
@@ -40,18 +40,6 @@ enum SpringMessagingAttributesGetter
 
   @Override
   @Nullable
-  public String getProtocol(MessageWithChannel messageWithChannel) {
-    return null;
-  }
-
-  @Override
-  @Nullable
-  public String getProtocolVersion(MessageWithChannel messageWithChannel) {
-    return null;
-  }
-
-  @Override
-  @Nullable
   public String getConversationId(MessageWithChannel messageWithChannel) {
     return null;
   }
