+++ instrumentation/spring/spring-rabbit-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/spring/rabbit/v1_0/SpringRabbitMessageAttributesGetter.java
@@ -37,18 +37,6 @@ enum SpringRabbitMessageAttributesGetter implements MessagingAttributesGetter<Me
 
   @Override
   @Nullable
-  public String getProtocol(Message message) {
-    return null;
-  }
-
-  @Override
-  @Nullable
-  public String getProtocolVersion(Message message) {
-    return null;
-  }
-
-  @Override
-  @Nullable
   public String getConversationId(Message message) {
     return null;
   }
